use lr in config file for transformer_tts, remove train.py in plain loop
